Added authentication and authorization with jwt

+++ routers/posts.py
@@ -6,6 +6,7 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse,  PostUpdate
+from auth import CurrentUser
 
 router = APIRouter()
 
@@ -18,15 +19,11 @@ async def get_posts(db:Annotated[AsyncSession, Depends(get_db)]):
     return posts
 
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post:PostCreate, db:Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).options(selectinload(models.User.posts)).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="user with the provided user_id does not exist")
+async def create_post(post:PostCreate, current_user:CurrentUser, db:Annotated[AsyncSession, Depends(get_db)]):
     new_post = models.Post(
         title = post.title,
         content = post.content,
-        user_id = post.user_id
+        user_id = current_user.id
     )
     db.add(new_post)
     await db.commit()
@@ -44,33 +41,33 @@ async def get_post(id: int, db:Annotated[AsyncSession, Depends(get_db)]):
     
 
 @router.put("/{id}", response_model=PostResponse)
-async def update_post_full(id: int, newPost:PostCreate, db:Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(id: int, newPost:PostCreate,current_user:CurrentUser, db:Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="you are not the owner of this post")
 
-    if newPost.user_id == post.user_id:
-        result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="user was not found")
-        
-        post.title = newPost.title
-        post.content = newPost.content
+    post.title = newPost.title
+    post.content = newPost.content
 
-        await db.commit()
-        await db.refresh(post, attribute_names=["author"])
-        return post
+    await db.commit()
+    await db.refresh(post, attribute_names=["author"])
+    return post
     
 
 @router.patch("/{id}", response_model=PostResponse)
-async def update_post_partial(id: int, newPost:PostUpdate, db:Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(id: int, newPost:PostUpdate, current_user:CurrentUser, db:Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
 
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="you are not the owner of this post")
+    
     newData = newPost.model_dump(exclude_unset=True)
     for field, value in newData.items():
         setattr(post, field, value)
@@ -80,12 +77,15 @@ async def update_post_partial(id: int, newPost:PostUpdate, db:Annotated[AsyncSes
     return post
     
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(id:int, db:Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(id:int,current_user:CurrentUser, db:Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="post not found")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="you are not the owner of this post")
 
     await db.delete(post)
     await db.commit()
